feat(frontend): scaffold ui design-system structure
- move existing reusable ui components into atomic design folders
- add atoms, molecules, organisms, and templates barrel exports
- keep packages/frontend/src/components/ui as the stable entry point
- add flat re-export shims so existing deep imports continue to resolve
- export component prop types from the reorganized design-system components
- update docs/component-refactor-plan.md with implementation notes
- mark Ticket 1 complete in the progress tracker
