+++ packages/frontend/src/components/ui/atoms/LoadingSpinner/index.ts
@@ -1,0 +1,2 @@
+export { LoadingSpinner } from './LoadingSpinner';
+export type { LoadingSpinnerProps } from './LoadingSpinner';
