+++ packages/frontend/src/components/ui/atoms/QuroLogo/index.ts
@@ -1,0 +1,2 @@
+export { QuroLogo } from './QuroLogo';
+export type { QuroLogoProps } from './QuroLogo';
