+++ packages/frontend/src/components/ui/atoms/index.ts
@@ -1,0 +1,2 @@
+export * from './LoadingSpinner';
+export * from './QuroLogo';
