+++ packages/frontend/src/components/ui/molecules/DateNoteRow/index.ts
@@ -1,0 +1,2 @@
+export { DateNoteRow } from './DateNoteRow';
+export type { DateNoteRowProps } from './DateNoteRow';
