+++ packages/frontend/src/components/ui/molecules/EmojiPickerField/index.ts
@@ -1,0 +1,2 @@
+export { EmojiPickerField } from './EmojiPickerField';
+export type { EmojiPickerFieldProps } from './EmojiPickerField';
