+++ packages/frontend/src/components/ui/molecules/EmptyState/index.ts
@@ -1,0 +1,2 @@
+export { EmptyState } from './EmptyState';
+export type { EmptyStateProps } from './EmptyState';
