+++ packages/frontend/src/components/ui/molecules/FormField/index.ts
@@ -1,0 +1,7 @@
+export { FormField, TextInput, CurrencyInput, SelectInput } from './FormField';
+export type {
+  CurrencyInputProps,
+  FormFieldProps,
+  SelectInputProps,
+  TextInputProps,
+} from './FormField';
