+++ packages/frontend/src/components/ui/molecules/PdfAttachmentField/index.ts
@@ -1,0 +1,2 @@
+export { PdfAttachmentField } from './PdfAttachmentField';
+export type { PdfAttachmentFieldProps } from './PdfAttachmentField';
