+++ packages/frontend/src/components/ui/molecules/TxnTypeSelector/index.ts
@@ -1,0 +1,2 @@
+export { TxnTypeSelector } from './TxnTypeSelector';
+export type { TxnTypeMeta, TxnTypeSelectorProps } from './TxnTypeSelector';
