+++ packages/frontend/src/components/ui/organisms/AreaChartCard/index.ts
@@ -1,0 +1,2 @@
+export { AreaChartCard } from './AreaChartCard';
+export type { AreaChartCardProps } from './AreaChartCard';
