+++ packages/frontend/src/components/ui/organisms/Modal/index.ts
@@ -1,0 +1,2 @@
+export { Modal, ModalFooter } from './Modal';
+export type { ModalFooterProps, ModalProps } from './Modal';
