+++ packages/frontend/src/components/ui/organisms/StatCard/index.ts
@@ -1,0 +1,2 @@
+export { StatCard } from './StatCard';
+export type { StatCardProps } from './StatCard';
