+++ packages/frontend/src/components/ui/organisms/TxnHistoryPanel/index.ts
@@ -1,0 +1,7 @@
+export { TxnHistoryPanel, TxnRow } from './TxnHistoryPanel';
+export type {
+  FilterOption,
+  TxnHistoryPanelProps,
+  TxnHistoryStat,
+  TxnRowProps,
+} from './TxnHistoryPanel';
