+++ packages/frontend/src/components/ui/templates/index.ts
@@ -1,0 +1,1 @@
+export {};
